--Added run_control struct; removed many local parameters from run_gryfx; fixed a bug which meant that nlpm->Phi_zf_kx1_avg was being read incorrectly; all tests pass
git-svn-id: svn://gk.umd.edu/noah/branches/gryfx_march15@2623 28a56f14-469b-1442-acfb-e68ebda8abd8

+++ diagnostics.cu
@@ -1288,8 +1288,8 @@ inline void restartWrite(cuComplex** Dens, cuComplex** Upar, cuComplex** Tpar, c
   cudaMemcpy(zCorr_sum_h, zCorr_sum, sizeof(float)*(Ny/2+1)*Nz, cudaMemcpyDeviceToHost);
   
   fwrite(&counter,sizeof(int),1,restart);
-  fwrite(&runtime,sizeof(float),1,restart);
-  fwrite(&dt,sizeof(float),1,restart);
+  fwrite(&runtime,sizeof(double),1,restart);
+  fwrite(&dt,sizeof(double),1,restart);
   fwrite(&timer,sizeof(float),1,restart);
   
   fwrite(wpfxAvg, sizeof(float)*nSpecies, 1, restart);
@@ -1364,8 +1364,8 @@ inline void restartRead(everything_struct * ev_h, everything_struct * ev_hd, flo
   zCorr_sum_h = (float*) malloc(sizeof(float)*(Ny/2+1)*Nz);
     
   fread(&ev_h->time.counter,sizeof(int),1,restart);  
-  fread(&ev_h->time.runtime,sizeof(float),1,restart); 
-  fread(&ev_h->time.dt, sizeof(float),1,restart);
+  fread(&ev_h->time.runtime,sizeof(double),1,restart); 
+  fread(&ev_h->time.dt, sizeof(double),1,restart);
   fread(&ev_h->time.totaltimer, sizeof(float),1,restart);
   /* Update global*/
   dt = ev_h->time.dt;
